face-detection/requirements.txt
click==8.1.3
cmake==3.24.2
colorama==0.4.6
dlib==19.24.0
face-recognition-models==0.3.0
face-recognition==1.3.0
future==0.18.2
imutils==0.5.4
iso8601==1.1.0
lewansoul-lx16a==0.1
numpy==1.23.4
opencv-python==4.6.0.66
paho-mqtt==1.6.1
picamera==1.13
Pillow==9.3.0
pyserial==3.5
PyYAML==6.0
state-machine==0.2.10